requirements.txt
dj-database-url==0.4.2
Django==1.11.4
gunicorn==19.7.1
olefile==0.44
Pillow==4.2.1
pytz==2017.2
whitenoise==3.3.0
psycopg2==2.6.2
django-autoslug==1.9.3

